refactor(base): split _base.py into _base_channel + _base_ion (ARCH-03)
- _base_channel.py: IonChannel, Channel, Synapse, IonInfo
- _base_ion.py: Ion, MixIons, mix_ions (root_type bound at end after late
  HHTypedNeuron import to avoid circular load)
- _base.py: keeps HHTypedNeuron plus re-exports. Every
  'from braincell._base import X' keeps working.
