+++ braincell/_base.py
@@ -417,940 +417,20 @@ class HHTypedNeuron(brainpy.state.Dynamics, Container, DiffEqModule):
             * self.spk_fun((V_th - last_V) / denom)
         )
 
-
-class IonChannel(brainstate.graph.Node, TreeNode, DiffEqModule):
-    """
-    Base class for modeling ion channel dynamics in neuronal simulations.
-
-    The IonChannel class serves as a foundation for implementing various types of ion channels,
-    including those for specific ions (e.g., sodium, potassium) or mixtures of ions. It provides
-    a structure for defining the behavior and properties of ion channels within a neuron model.
-
-    This class is designed to be subclassed to create specific ion channel models. Subclasses
-    should implement the core methods to define the channel's behavior, such as current calculation,
-    state initialization, and derivative computation.
-
-    Attributes
-    ----------
-    in_size : tuple
-        The dimensions of the ion channel, representing its size (e.g., number of neurons,
-        number of compartments).
-    out_size : tuple
-        Same as in_size, representing the output dimensions of the channel.
-    name : str, optional
-        A name identifier for the ion channel.
-
-    Notes
-    -----
-    - Subclasses should override the abstract methods (current, compute_derivative, init_state,
-      reset_state) to define the specific behavior of the ion channel.
-    - The class integrates with the broader neuron modeling framework, allowing for complex
-      simulations of neuronal dynamics.
-    - It's designed to work within a hierarchical structure of neuronal components, as indicated
-      by its inheritance from TreeNode.
-
-    Example
-    -------
-
-    .. code-block:: python
-
-        class SodiumChannel(IonChannel):
-            def __init__(self, size, g_max):
-                super().__init__(size)
-                self.g_max = g_max
-
-            def current(self, V, Na):
-                # Implement sodium current calculation
-                pass
-
-            def compute_derivative(self, V, Na):
-                # Implement derivative computation for channel states
-                pass
-
-            def init_state(self, V, Na, batch_size=None):
-                # Initialize channel states
-                pass
-
-            def reset_state(self, V, Na, batch_size=None):
-                # Reset channel states
-                pass
-    """
-
-    __module__ = 'braincell'
-
-    def __init__(
-        self,
-        size: brainstate.typing.Size,
-        name: Optional[str] = None,
-    ):
-        # size
-        if isinstance(size, (list, tuple)):
-            if len(size) <= 0:
-                raise ValueError(f'size must be int, or a tuple/list of int. '
-                                 f'But we got {type(size)}')
-            if not isinstance(size[0], (int, np.integer)):
-                raise ValueError('size must be int, or a tuple/list of int.'
-                                 f'But we got {type(size)}')
-            size = tuple(size)
-        elif isinstance(size, (int, np.integer)):
-            size = (size,)
-        else:
-            raise ValueError('size must be int, or a tuple/list of int.'
-                             f'But we got {type(size)}')
-        self.size = size
-        assert len(size) >= 1, ('The size of the dendritic dynamics should be at '
-                                'least 1D: (..., n_neuron, n_compartment).')
-        self.name = name
-
-    @property
-    def varshape(self):
-        """
-        Get the shape of variables in the neuron group.
-
-        Returns
-        -------
-        tuple
-            The shape of variables, typically representing the dimensions of the neuron group.
-        """
-        return self.size
-
-    def current(self, *args, **kwargs):
-        """
-        Calculate the current for this ion channel.
-
-        This method should be implemented by subclasses to compute the current
-        based on the channel's specific properties and state.
-
-        Parameters
-        ----------
-        *args
-            Variable length argument list.
-        **kwargs
-            Arbitrary keyword arguments.
-
-        Raises
-        ------
-        NotImplementedError
-            This method must be implemented by subclasses.
-        """
-        raise NotImplementedError
-
-    def pre_integral(self, *args, **kwargs):
-        """
-        Perform pre-integration operations.
-
-        This method is called before the integration step in simulations.
-        It can be used to prepare the channel's state or perform any necessary
-        calculations before integration.
-
-        Parameters
-        ----------
-        *args
-            Variable length argument list.
-        **kwargs
-            Arbitrary keyword arguments.
-        """
-        pass
-
-    def compute_derivative(self, *args, **kwargs):
-        """
-        Compute the derivative of the channel's state variables.
-
-        This method should be implemented by subclasses to calculate how the
-        channel's state changes over time.
-
-        Parameters
-        ----------
-        *args
-            Variable length argument list.
-        **kwargs
-            Arbitrary keyword arguments.
-
-        Raises
-        ------
-        NotImplementedError
-            This method must be implemented by subclasses.
-        """
-        raise NotImplementedError
-
-    def post_integral(self, *args, **kwargs):
-        """
-        Perform post-integration operations.
-
-        This method is called after the integration step in simulations.
-        It should be used to update the channel's state based on the results of integration.
-
-        Parameters
-        ----------
-        *args
-            Variable length argument list.
-        **kwargs
-            Arbitrary keyword arguments.
-
-        Raises
-        ------
-        NotImplementedError
-            This method must be implemented by subclasses.
-        """
-        pass
-
-    def reset_state(self, *args, **kwargs):
-        """
-        Reset the state of the ion channel.
-
-        This method should reset all state variables of the channel to their initial values.
-
-        Parameters
-        ----------
-        *args
-            Variable length argument list.
-        **kwargs
-            Arbitrary keyword arguments.
-        """
-        pass
-
-    def init_state(self, *args, **kwargs):
-        """
-        Initialize the state of the ion channel.
-
-        This method should set up the initial state of all variables for the channel.
-
-        Parameters
-        ----------
-        *args
-            Variable length argument list.
-        **kwargs
-            Arbitrary keyword arguments.
-        """
-        pass
-
-    def update(self, *args, **kwargs):
-        if isinstance(self, IndependentIntegration):
-            self.make_integration(*args, **kwargs)
-
-    def _on_param_updated(self, var_name: str, new_value) -> None:
-        """Hook invoked after runtime state writes a parameter.
-
-        Default: no-op. Subclasses override to recompute derived
-        values when a specific parameter changes (for example, HH1952
-        channels recompute ``phi`` when ``T`` changes).
-
-        Parameters
-        ----------
-        var_name : str
-            The parameter name that was just updated.
-        new_value : object
-            The new value written to the runtime node attribute.
-            Typically a :class:`brainunit.Quantity` but may be a plain
-            array.
-        """
-        return None
-
-
-class IonInfo(NamedTuple):
-    """
-    A named tuple representing the information of an ion in a neuron model.
-
-    This class encapsulates the intracellular/extracellular concentrations of
-    an ion, its reversal potential, and its valence. It is used to store and
-    pass ion-related information in various neuronal simulation contexts.
-
-    Attributes:
-        Ci (brainstate.typing.ArrayLike): The intracellular ion concentration.
-            This represents the concentration of the ion inside the cell,
-            typically in units of millimoles per liter (mM).
-
-        Co (brainstate.typing.ArrayLike): The extracellular ion concentration.
-            This represents the concentration of the ion outside the cell,
-            typically in units of millimoles per liter (mM).
-
-        E (brainstate.typing.ArrayLike): The reversal potential.
-            This represents the electrical potential at which there is no net
-            flow of the ion across the membrane, typically in millivolts (mV).
-
-        valence (brainstate.typing.ArrayLike): The ionic valence.
-            This represents the charge number used in Nernst/GHK relations.
-
-    Note:
-        ``Ci``, ``Co``, ``E``, and ``valence`` are expected to be array-like
-        objects or scalars, allowing representation of these properties across
-        multiple neurons or compartments simultaneously.
-    """
-    Ci: brainstate.typing.ArrayLike
-    Co: brainstate.typing.ArrayLike
-    E: brainstate.typing.ArrayLike
-    valence: brainstate.typing.ArrayLike
-
-
-class Ion(IonChannel, Container):
-    """
-    The base class for modeling ion dynamics in neuronal simulations.
-
-    This class represents a specific type of ion (e.g., sodium, potassium) and manages
-    the associated ion channels and their dynamics. It inherits from both IonChannel
-    and Container, allowing it to handle ion-specific behaviors and contain multiple
-    channel instances.
-
-    Args:
-        size (brainstate.typing.Size): The size of the simulation target, typically
-            representing the number of neurons or compartments.
-        name (Optional[str]): The name of the Ion instance. Defaults to None.
-        **channels: Additional keyword arguments for specifying Channel instances
-            to be included in this Ion object.
-
-    Attributes:
-        channels (Dict[str, Channel]): A dictionary of Channel instances associated
-            with this ion.
-
-    The Ion class serves as a crucial component in modeling the behavior of specific
-    ion types within a neuron or neural network simulation. It manages the collective
-    behavior of multiple ion channels of the same ion type and provides methods for
-    initializing, updating, and querying the state of these channels throughout the
-    simulation process.
-
-    Parameters:
-    -----------
-    size : brainstate.typing.Size
-        The size of the ion channel, typically representing the number of
-        neurons or compartments.
-    name : Optional[str], default=None
-        The name of the Ion instance. If not provided, the instance will be unnamed.
-    **channels
-        Additional keyword arguments for specifying Channel instances to be
-        included in this Ion object.
-    """
-
-    __module__ = 'braincell'
-    _container_name = 'channels'
-
-    # The type of the master object.
-    root_type = HHTypedNeuron
-
-    def __init__(
-        self,
-        size: brainstate.typing.Size,
-        name: Optional[str] = None,
-        **channels
-    ) -> None:
-        super().__init__(size, name=name)
-        self.channels: Dict[str, Channel] = dict()
-        self.channels.update(self._format_elements(Channel, **channels))
-
-        self._external_currents: Dict[str, Callable] = dict()
-
-    @property
-    def external_currents(self) -> Dict[str, Callable]:
-        """
-        Get the dictionary of external currents.
-
-        Returns:
-            Dict[str, Callable]: A dictionary where keys are strings identifying the external currents,
-                                 and values are callable functions representing those currents.
-        """
-        return self._external_currents
-
-    def pre_integral(self, V):
-        """
-        Perform pre-integration operations for all channels.
-
-        This method is called before the integration step in simulations. It iterates through
-        all Channel nodes and calls their pre_integral methods.
-
-        Parameters:
-            V (array-like): The membrane potential for all neurons/compartments.
-        """
-        nodes = brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1))
-        for node in nodes.values():
-            if not isinstance(node, IndependentIntegration):
-                node.pre_integral(V, self.pack_info())
-
-    def compute_derivative(self, V):
-        """
-        Compute derivatives for all channels.
-
-        This method calculates the derivatives of state variables for all Channel nodes.
-
-        Parameters:
-            V (array-like): The membrane potential for all neurons/compartments.
-        """
-        nodes = tuple(brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).values())
-        self.check_hierarchies(type(self), *nodes)
-        ion_info = self.pack_info()
-        for node in nodes:
-            if not isinstance(node, IndependentIntegration):
-                node.compute_derivative(V, ion_info)
-        self._run_ion_hook("_ion_compute_derivative_hook", V)
-
-    def post_integral(self, V):
-        """
-        Perform post-integration operations for all channels.
-
-        This method is called after the integration step in simulations. It iterates through
-        all Channel nodes and calls their post_integral methods.
-
-        Parameters:
-            V (array-like): The membrane potential for all neurons/compartments.
-        """
-        nodes = brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1))
-        for node in nodes.values():
-            if not isinstance(node, IndependentIntegration):
-                node.post_integral(V, self.pack_info())
-
-    def current(self, V, include_external: bool = False):
-        """
-        Generate ion channel current.
-
-        This method calculates the total current from all channels and optionally includes external currents.
-
-        Parameters:
-            V (array-like): The membrane potential for all neurons/compartments.
-            include_external (bool): If True, include external currents in the calculation. Default is False.
-
-        Returns:
-            array-like: The total current generated by all channels (and external currents if included).
-        """
-        nodes = tuple(brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).values())
-
-        ion_info = self.pack_info()
-        current = None
-        if len(nodes) > 0:
-            for node in nodes:
-                node: Channel
-                new_current = node.current(V, ion_info)
-                current = new_current if current is None else (current + new_current)
-        if include_external and self._external_currents:
-            for key, node in self._external_currents.items():
-                node: Callable
-                contrib = node(V, ion_info)
-                current = contrib if current is None else (current + contrib)
-        return current
-
-    def init_state(self, V, batch_size: int = None):
-        """
-        Initialize the state of all channels.
-
-        This method initializes the state variables for all Channel nodes.
-
-        Parameters:
-            V (array-like): The membrane potential for all neurons/compartments.
-            batch_size (int, optional): The batch size for initialization. Default is None.
-        """
-        nodes = tuple(brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).values())
-        self.check_hierarchies(type(self), *nodes)
-        self._run_ion_hook("_ion_init_state_hook", V, batch_size=batch_size)
-        ion_info = self.pack_info()
-        for node in nodes:
-            node: Channel
-            node.init_state(V, ion_info, batch_size)
-
-    def reset_state(self, V, batch_size: int = None):
-        """
-        Reset the state of all channels.
-
-        This method resets the state variables for all Channel nodes to their initial values.
-
-        Parameters:
-            V (array-like): The membrane potential for all neurons/compartments.
-            batch_size (int, optional): The batch size for resetting. Default is None.
-        """
-        nodes = tuple(brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).values())
-        self.check_hierarchies(type(self), *nodes)
-        self._run_ion_hook("_ion_reset_state_hook", V, batch_size=batch_size)
-        ion_info = self.pack_info()
-        for node in nodes:
-            node: Channel
-            node.reset_state(V, ion_info, batch_size)
-
-    def update(self, V, *args, **kwargs):
-        ion_info = self.pack_info()
-        for key, node in brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).items():
-            node.update(V, ion_info)
-
-    def _run_ion_hook(self, name: str, *args, **kwargs):
-        hook = getattr(self, name, None)
-        if hook is not None:
-            hook(*args, **kwargs)
-
-    def register_external_current(self, key: Hashable, fun: Callable):
-        """
-        Register an external current function.
-
-        This method adds a new external current function to the ion channel.
-
-        Parameters:
-            key (Hashable): A unique identifier for the external current.
-            fun (Callable): The function that computes the external current.
-
-        Raises:
-            ValueError: If the key already exists in the external currents' dictionary.
-        """
-        if key in self._external_currents:
-            raise ValueError
-        self._external_currents[key] = fun
-
-    def pack_info(self) -> IonInfo:
-        """
-        Pack the ion information into an IonInfo object.
-
-        This method collects the intracellular/extracellular concentrations
-        (Ci/Co), reversal potential (E), and valence of the ion and packages
-        them into an IonInfo named tuple.
-
-        Returns:
-            IonInfo: A named tuple containing:
-
-                - Ci (array-like): The intracellular ion concentration.
-                - Co (array-like): The extracellular ion concentration.
-                - E (array-like): The reversal potential of the ion.
-                - valence (array-like): The ionic valence.
-
-        Notes:
-            If an attribute is an instance of brainstate.State, its ``value``
-            attribute is used. Otherwise, the raw attribute value is used.
-        """
-        Ci = self.Ci
-        Ci = Ci.value if isinstance(Ci, brainstate.State) else Ci
-        Co = self.Co
-        Co = Co.value if isinstance(Co, brainstate.State) else Co
-        E = self.E
-        E = E.value if isinstance(E, brainstate.State) else E
-        valence = self.valence
-        valence = valence.value if isinstance(valence, brainstate.State) else valence
-        return IonInfo(Ci=Ci, Co=Co, E=E, valence=valence)
-
-    def add(self, **elements):
-        """
-        Add new channel elements to the Ion instance.
-
-        This method adds new Channel instances to the Ion object. It checks the
-        hierarchies of the new elements and updates the channels dictionary.
-
-        Parameters
-        ----------
-        **elements : Any
-            A dictionary of new elements to add. Each key-value pair represents a
-            channel name and its corresponding Channel instance.
-
-        Raises
-        ------
-        TypeError
-            If the hierarchies of the new elements are incompatible with the current structure.
-
-        Notes
-        -----
-        - The method checks hierarchies using the check_hierarchies method.
-        - New elements are formatted and added to the channels dictionary.
-        """
-        self.check_hierarchies(type(self), **elements)
-        self.channels.update(self._format_elements(Channel, **elements))
-
-
-class MixIons(IonChannel, Container):
-    """
-    A class for mixing multiple ion channels in neuronal simulations.
-
-    This class combines multiple Ion instances to create a composite ion channel
-    that can handle the dynamics of multiple ion types simultaneously.
-
-    Args:
-        *ions: Variable number of Ion instances. These define the types of ions
-               that will be mixed in this channel.
-        name (Optional[str]): The name of the MixIons instance. Defaults to None.
-        **channels: Additional keyword arguments for specifying Channel instances.
-
-    Attributes:
-        ions (Sequence['Ion']): A tuple of Ion instances that are part of this mixed channel.
-        ion_types (tuple): A tuple of the types of the Ion instances.
-        channels (Dict[str, Channel]): A dictionary of Channel instances associated with this mixed channel.
-
-    Raises:
-        AssertionError: If fewer than two ions are provided, if any provided ion is not an Ion instance,
-                        or if the sizes of all provided ions are not identical.
-    """
-    __module__ = 'braincell'
-
-    root_type = HHTypedNeuron
-    _container_name = 'channels'
-
-    def __init__(self, *ions, name: Optional[str] = None, **channels):
-        """
-        Initialize a MixIons instance for combining multiple ion channels.
-
-        This constructor creates a MixIons object that combines multiple Ion instances
-        to model the behavior of a composite ion channel. It ensures that at least two
-        ions are provided and that all ions have consistent sizes.
-
-        Parameters
-        ----------
-        *ions
-            Variable number of Ion instances to be combined in this mixed channel.
-            At least two ions must be provided.
-
-        name : Optional[str], default=None
-            The name of this MixIons instance. If not provided, the instance will be unnamed.
-
-        **channels
-            Additional keyword arguments for specifying Channel instances to be included
-            in this MixIons object.
-
-        Raises
-        ------
-        AssertionError
-            If fewer than two ions are provided, if any provided ion is not an Ion instance,
-            or if the sizes of all provided ions are not identical.
-
-        Notes
-        -----
-        - The size of the MixIons instance is set to the size of the first provided ion.
-        - All provided ions must have the same size.
-        - The method stores both the ion instances and their types.
-        - Additional channels can be added via keyword arguments.
-        """
-        # TODO: check "ion" should be independent from each other
-        assert len(ions) >= 2, f'{self.__class__.__name__} requires at least two ion. '
-        assert all([isinstance(cls, Ion) for cls in ions]), f'Must be a sequence of Ion. But got {ions}.'
-        size = ions[0].size
-        for ion in ions:
-            assert ion.size == size, f'The size of all ion should be the same. But we got {ions}.'
-        super().__init__(size=size, name=name)
-
-        # Store the ion instances
-        self.ions: Sequence['Ion'] = tuple(ions)
-        self._ion_types = tuple([type(ion) for ion in self.ions])
-
-        # Store the ion channel channel
-        self.channels: Dict[str, Channel] = dict()
-        self.channels.update(self._format_elements(Channel, **channels))
-
-    @property
-    def ion_types(self) -> Tuple[Type[Ion], ...]:
-        """
-        Get the types of ions in this mixed channel.
-
-        Returns:
-            Tuple[Type[Ion], ...]: A tuple containing the types of all Ion instances in this mixed channel.
-        """
-        return self._ion_types
-
-    def pre_integral(self, V):
-        """
-        Perform pre-integration operations for all channels in the mixed ion group.
-
-        This method is called before the integration step in the simulation. It iterates
-        through all Channel nodes in the MixIons instance and calls their pre_integral
-        methods with the current membrane potential and relevant ion information.
-
-        Parameters
-        ----------
-        V : array-like
-            The current membrane potential for all neurons in the simulation.
-
-        Notes
-        -----
-        The method retrieves all Channel nodes, packs the necessary ion information
-        for each channel, and then calls the pre_integral method of each channel
-        with the current membrane potential and the packed ion information.
-        """
-        nodes = tuple(brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).values())
-        for node in nodes:
-            if not isinstance(node, IndependentIntegration):
-                ion_infos = tuple([self._get_ion(ion).pack_info() for ion in node.root_type.__args__])
-                node.pre_integral(V, *ion_infos)
-
-    def compute_derivative(self, V):
-        """
-        Compute the derivative of state variables for all channels in the mixed ion group.
-
-        This method iterates through all Channel nodes in the MixIons instance and calls
-        their compute_derivative methods with the current membrane potential and relevant
-        ion information. It's typically used in the process of numerically solving the
-        differential equations that describe the ion channel dynamics.
-
-        Parameters
-        ----------
-        V : array-like
-            The current membrane potential for all neurons in the simulation.
-
-        Notes:
-        ------
-        The method retrieves all Channel nodes, packs the necessary ion information
-        for each channel, and then calls the compute_derivative method of each channel
-        with the current membrane potential and the packed ion information.
-        """
-        nodes = tuple(brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).values())
-        for node in nodes:
-            if not isinstance(node, IndependentIntegration):
-                ion_infos = tuple([self._get_ion(ion).pack_info() for ion in node.root_type.__args__])
-                node.compute_derivative(V, *ion_infos)
-
-    def post_integral(self, V):
-        """
-        Perform post-integration operations for all channels in the mixed ion group.
-
-        This method is called after the integration step in the simulation. It iterates
-        through all Channel nodes in the MixIons instance and calls their post_integral
-        methods with the current membrane potential and relevant ion information.
-
-        Parameters
-        ----------
-        V : array-like
-            The current membrane potential for all neurons in the simulation.
-
-        Notes
-        -----
-        The method retrieves all Channel nodes, packs the necessary ion information
-        for each channel, and then calls the post_integral method of each channel
-        with the current membrane potential and the packed ion information.
-        """
-        nodes = tuple(brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).values())
-        for node in nodes:
-            if not isinstance(node, IndependentIntegration):
-                ion_infos = tuple([self._get_ion(ion).pack_info() for ion in node.root_type.__args__])
-                node.post_integral(V, *ion_infos)
-
-    def current(self, V):
-        """
-        Generate the total ion channel current for all channels in the mixed ion group.
-
-        This method calculates the sum of currents from all Channel nodes within the
-        MixIons instance. It iterates through each channel, computes its individual
-        current, and aggregates them.
-
-        Parameters
-        ----------
-        V : array-like
-            The membrane potential for all neurons in the simulation. This is typically
-            a numpy array or similar structure containing voltage values.
-
-        Returns
-        -------
-        array-like or float
-            The total current generated by all channels. If there are no channels,
-            it returns 0. Otherwise, it returns an array-like object of the same
-            shape as V, representing the summed current at each point.
-
-        Notes
-        -----
-        - The method retrieves all Channel nodes and their corresponding ion information.
-        - For each channel, it calls the channel's current method with the membrane
-          potential and relevant ion information.
-        - The currents from all channels are summed to produce the total current.
-        """
-        nodes = tuple(brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).values())
-
-        if len(nodes) == 0:
-            return 0.
-        else:
-            current = None
-            for node in nodes:
-                infos = tuple([self._get_ion(root).pack_info() for root in node.root_type.__args__])
-                current = (
-                    node.current(V, *infos)
-                    if current is None else
-                    (current + node.current(V, *infos))
-                )
-            return current
-
-    def init_state(self, V, batch_size: int = None):
-        """
-        Initialize the state of all channels in the mixed ion group.
-
-        This method iterates through all Channel nodes in the MixIons instance and initializes
-        their states. It checks the hierarchies of the nodes and packs the necessary ion
-        information for each channel before calling their init_state methods.
-
-        Parameters
-        ----------
-        V : array-like
-            The initial membrane potential for all neurons in the simulation.
-        batch_size : int, optional
-            The batch size for the simulation. If provided, it will be used to initialize
-            the states with the specified batch size.
-
-        Notes
-        -----
-        The method retrieves all Channel nodes, checks their hierarchies, packs the necessary
-        ion information for each channel, and then calls the init_state method of each channel
-        with the initial membrane potential, ion information, and batch size.
-        """
-        nodes = brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).values()
-        self.check_hierarchies(self.ion_types, *tuple(nodes), check_fun=self._check_hierarchy)
-        for node in nodes:
-            node: Channel
-            infos = tuple([self._get_ion(root).pack_info() for root in node.root_type.__args__])
-            node.init_state(V, *infos, batch_size)
-
-    def reset_state(self, V, batch_size=None):
-        """
-        Reset the state of all channels in the mixed ion group.
-
-        This method iterates through all Channel nodes in the MixIons instance and resets
-        their states to initial values. It's typically used when restarting a simulation
-        or when transitioning between different phases of a simulation.
-
-        Parameters
-        ----------
-        V : array-like
-            The current membrane potential for all neurons in the simulation.
-            This is used to reset the channel states based on the current voltage.
-
-        batch_size : int, optional
-            The batch size for the simulation. If provided, it will be used to reset
-            the states with the specified batch size, ensuring consistency in batched simulations.
-
-        Notes
-        -----
-        The method retrieves all Channel nodes, packs the necessary ion information
-        for each channel, and then calls the reset_state method of each channel
-        with the current membrane potential, ion information, and batch size.
-        """
-        nodes = tuple(brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).values())
-        for node in nodes:
-            infos = tuple([self._get_ion(root).pack_info() for root in node.root_type.__args__])
-            node.reset_state(V, *infos, batch_size)
-
-    def update(self, V, *args, **kwargs):
-        for key, node in brainstate.graph.nodes(self, Channel, allowed_hierarchy=(1, 1)).items():
-            infos = tuple([self._get_ion(root).pack_info() for root in node.root_type.__args__])
-            node.update(V, *infos)
-
-    def _check_hierarchy(self, ions, leaf):
-        # 'root_type' should be a brainpy.mixin.JointType
-        self._check_root(leaf)
-        for cls in leaf.root_type.__args__:
-            if not any([issubclass(root, cls) for root in ions]):
-                raise TypeError(
-                    f'Type does not match. {leaf} requires a master with type '
-                    f'of {leaf.root_type}, but the master type now is {ions}.'
-                )
-
-    def add(self, **elements):
-        """
-        Add new channel elements to the MixIons instance.
-
-        This method adds new Channel instances to the MixIons object. It checks the
-        hierarchies of the new elements, updates the channels dictionary, and registers
-        external currents for each new channel with its corresponding ion.
-
-        Parameters
-        ----------
-        **elements
-            A dictionary of new elements to add. Each key-value pair represents a
-            channel name and its corresponding Channel instance.
-
-        Raises
-        ------
-        TypeError
-            If the hierarchies of the new elements are incompatible with the current structure.
-
-        Notes
-        -----
-        - The method checks hierarchies using the _check_hierarchy method.
-        - It updates the channels dictionary with formatted Channel elements.
-        - For each new channel, it registers an external current with each associated ion.
-        """
-        self.check_hierarchies(self.ion_types, check_fun=self._check_hierarchy, **elements)
-        self.channels.update(self._format_elements(Channel, **elements))
-        for elem in tuple(elements.values()):
-            elem: Channel
-            for ion_root in elem.root_type.__args__:
-                ion = self._get_ion(ion_root)
-                ion.register_external_current(id(elem), self._get_ion_fun(ion, elem))
-
-    def _get_ion_fun(self, ion: 'Ion', node: 'Channel'):
-        def fun(V, ion_info):
-            infos = tuple([
-                (ion_info if isinstance(ion, root) else self._get_ion(root).pack_info())
-                for root in node.root_type.__args__
-            ])
-            return node.current(V, *infos)
-
-        return fun
-
-    def _get_ion(self, cls):
-        for ion in self.ions:
-            if isinstance(ion, cls):
-                return ion
-        else:
-            raise ValueError(f'No instance of {cls} is found.')
-
-    def _check_root(self, leaf):
-        if not isinstance(leaf.root_type, _JointGenericAlias):
-            raise TypeError(
-                f'{self.__class__.__name__} requires leaf nodes that have the root_type of '
-                f'"brainpy.mixin.JointType". However, we got {leaf.root_type}'
-            )
-
-
-@set_module_as('braincell')
-def mix_ions(*ions) -> MixIons:
-    """
-    Create a mixed ion channel by combining multiple ion instances.
-
-    This function takes one or more Ion instances and creates a MixIons object,
-    which represents a channel that can handle multiple types of ions simultaneously.
-
-    Parameters
-    ----------
-    *ions
-        One or more instances of the Ion class. Each instance represents a specific
-        type of ion (e.g., sodium, potassium, calcium) that will be part of the
-        mixed ion channel.
-
-    Returns
-    -------
-    MixIons
-        An instance of the MixIons class that combines all the provided ion instances
-        into a single mixed ion channel.
-
-    Raises
-    ------
-    AssertionError
-        If no ions are provided or if any of the provided arguments is not an instance
-        of the Ion class.
-
-    Examples:
-    ---------
-    >>> import braincell    
-    >>> sodium_ion = braincell.ion.SodiumFixed(...)
-    >>> potassium_ion = braincell.ion.PotassiumFixed(...)
-    >>> mixed_channel = mix_ions(sodium_ion, potassium_ion)
-    """
-    for ion in ions:
-        assert isinstance(ion, Ion), f'Must be instance of {Ion.__name__}. But got {type(ion)}'
-    assert len(ions) >= 2, f'mix_ions requires at least two ions, got {len(ions)}.'
-    return MixIons(*ions)
-
-
-class Channel(IonChannel):
-    """
-    The base class for modeling channel dynamics in neuronal simulations.
-
-    This class extends the IonChannel class to provide a framework for implementing
-    specific ion channel models. It serves as a foundation for creating various types
-    of ion channels, such as voltage-gated or ligand-gated channels.
-
-    Note:
-        Subclasses of Channel should implement specific methods like `current`,
-        `compute_derivative`, etc., to define the behavior of the particular channel type.
-
-    Example::
-
-    .. code-block:: python
-
-        class SodiumChannel(Channel):
-            def __init__(self, *args, **kwargs):
-                super().__init__(*args, **kwargs)
-                # Additional initialization for sodium channel
-
-            def current(self, V, *args):
-                # Implement sodium current calculation
-                pass
-
-            # Implement other required methods
-    """
-    __module__ = 'braincell'
-
-
-class Synapse(IonChannel):
-    __module__ = 'braincell'
+# ---------------------------------------------------------------------------
+# Re-exports (ARCH-03). The ion-channel family lives in
+# ``braincell._base_channel``; ion species and the ``mix_ions`` factory
+# live in ``braincell._base_ion``. Every existing caller using
+# ``from braincell._base import X`` continues to work through these
+# re-exports.
+from ._base_channel import (  # noqa: E402
+    Channel,
+    IonChannel,
+    IonInfo,
+    Synapse,
+)
+from ._base_ion import (  # noqa: E402
+    Ion,
+    MixIons,
+    mix_ions,
+)
